+++ packages/plugins/eslint-plugin-react-dom/src/rules/no-hydrate.spec.ts
@@ -7,15 +7,15 @@ ruleTester.run(RULE_NAME, rule, {
   invalid: [
     {
       code: tsx`
-        import ReactDom from "react-dom";
+        import ReactDOM from "react-dom";
         import Component from "Component";
 
-        ReactDom.hydrate(<Component />, document.getElementById("app"));
+        ReactDOM.hydrate(<Component />, document.getElementById("app"));
       `,
       errors: [{ messageId: "noHydrate" }],
       output: tsx`
         import { hydrateRoot } from "react-dom/client";
-        import ReactDom from "react-dom";
+        import ReactDOM from "react-dom";
         import Component from "Component";
 
         hydrateRoot(document.getElementById("app"), <Component />);
@@ -24,16 +24,16 @@ ruleTester.run(RULE_NAME, rule, {
     {
       code: tsx`
         import React from "react";
-        import ReactDom from "react-dom";
+        import ReactDOM from "react-dom";
         import Component from "Component";
 
-        ReactDom.hydrate(<Component />, document.getElementById("app")!);
+        ReactDOM.hydrate(<Component />, document.getElementById("app")!);
       `,
       errors: [{ messageId: "noHydrate" }],
       output: tsx`
         import { hydrateRoot } from "react-dom/client";
         import React from "react";
-        import ReactDom from "react-dom";
+        import ReactDOM from "react-dom";
         import Component from "Component";
 
         hydrateRoot(document.getElementById("app")!, <Component />);
@@ -42,17 +42,17 @@ ruleTester.run(RULE_NAME, rule, {
     {
       code: tsx`
         import React from "react";
-        import ReactDom from "react-dom";
+        import ReactDOM from "react-dom";
         import Component from "Component";
 
         const rootEl = document.getElementById("app")!;
-        ReactDom.hydrate(<Component />, rootEl);
+        ReactDOM.hydrate(<Component />, rootEl);
       `,
       errors: [{ messageId: "noHydrate" }],
       output: tsx`
         import { hydrateRoot } from "react-dom/client";
         import React from "react";
-        import ReactDom from "react-dom";
+        import ReactDOM from "react-dom";
         import Component from "Component";
 
         const rootEl = document.getElementById("app")!;